Extract Playlist to its own file
