+++ src-tauri/src/libs/mod.rs
@@ -1,5 +1,8 @@
 pub mod error;
 pub mod events;
 pub mod file_associations;
-pub mod track;
 pub mod utils;
+
+// DB Structs
+pub mod playlist;
+pub mod track;
